+++ packages/core/test/isotest.js
@@ -24,7 +24,7 @@ const customChildren = {
     Popover2: popoverTarget,
     SVGPopover: popoverTarget,
     SVGTooltip: popoverTarget,
-    Tabs2: [Core.Tab2Factory({ key: 1, id: 1, title: "Tab one" })],
+    Tabs: React.createElement(Core.Tab, { key: 1, id: 1, title: "Tab one" }),
     Tooltip: popoverTarget,
     Tooltip2: popoverTarget,
 };
