+++ packages/docs-app/src/styles/_examples.scss
@@ -167,7 +167,7 @@
   }
 }
 
-.docs-tabs2-example .pt-navbar .pt-tab {
+.docs-tabs-example .pt-navbar .pt-tab {
   line-height: $pt-navbar-height;
 }
